Extract compression chart constants
