+++ src/app/components/charts/compression/compression-chart-builder.service.ts
@@ -4,6 +4,14 @@ import { formatBytes } from "../../../pipes/pipe-utils";
 import { radarDatasetColors } from "../shared/chart-colors.constants";
 import { cloneChartOptions } from "../shared/chart-options.utils";
 import {
+  COMPRESSION_LEVEL_NOT_AVAILABLE_LABEL,
+  DETAILS_SERIES_IGNORED_KEYS,
+  DetailsCompressionMode,
+  LEGACY_LZ4_BLOCK_SIZE_BYTES,
+  compressionInfoTooltip,
+  detailsCompressionModes,
+} from "./compression-chart.constants";
+import {
   CompareCompressionOption,
   CompressionBenchmarkGroup,
   CompressionBenchmarkMeta,
@@ -19,65 +27,11 @@ import {
   CompressionServer,
 } from "./compression-chart.types";
 
-export type DetailsCompressionMode = {
-  name: string;
-  key:
-    | "compress"
-    | "decompress"
-    | "ratio"
-    | "ratio_compress"
-    | "ratio_decompress";
-  order: string;
-  icon: "circle-arrow-down" | "circle-arrow-up";
-};
-
-const detailsCompressionModes: DetailsCompressionMode[] = [
-  {
-    name: "Compression speed",
-    key: "compress",
-    order: "Higher is better.",
-    icon: "circle-arrow-up",
-  },
-  {
-    name: "Decompression speed",
-    key: "decompress",
-    order: "Higher is better.",
-    icon: "circle-arrow-up",
-  },
-  {
-    name: "Compression ratio",
-    key: "ratio",
-    order: "Lower is better.",
-    icon: "circle-arrow-down",
-  },
-  {
-    name: "Compression speed/ratio",
-    key: "ratio_compress",
-    order: "Higher is better.",
-    icon: "circle-arrow-up",
-  },
-  {
-    name: "Decompression speed/ratio",
-    key: "ratio_decompress",
-    order: "Higher is better.",
-    icon: "circle-arrow-up",
-  },
-];
-
-const DETAILS_SERIES_IGNORED_KEYS = new Set(["algo", "compression_level", "cores"]);
-const COMPRESSION_LEVEL_NOT_AVAILABLE_LABEL = "N/A";
-const LEGACY_LZ4_BLOCK_SIZE_BYTES: Record<number, number> = {
-  0: 64 * 1024,
-  4: 1024 * 1024,
-  6: 4 * 1024 * 1024,
-};
-
 @Injectable({
   providedIn: "root",
 })
 export class CompressionChartBuilderService {
-  readonly infoTooltip =
-    "Measuring the compression ratio and speed while compressing and decompressing the dickens.txt of the Silesia corpus (10 MB uncompressed) using various algorithms, compressions levels and other extra arguments.";
+  readonly infoTooltip = compressionInfoTooltip;
 
   getDetailsModes(): DetailsCompressionMode[] {
     return detailsCompressionModes;
